Stop using apl-fy24 branch of repos

+++ src/cace/amm/typing.c
@@ -604,7 +604,7 @@ static bool builtin_time_constraints(struct timespec *out, const cace_ari_t *ari
         case CACE_ARI_PRIM_FLOAT64:
         {
             double v = obj->value.as_float64;
-            if (isnan(v) || isinf(v) || v > INT64_MAX || v < INT64_MIN)
+            if (isnan(v) || isinf(v) || v > (double)INT64_MAX || v < (double)INT64_MIN)
             {
                 return false;
             }
